+++ src/aks-preview/azext_aks_preview/custom.py
@@ -127,9 +127,7 @@ from azext_aks_preview.maintenanceconfiguration import (
 )
 
 from .vendored_sdks.azure_mgmt_preview_aks.v2022_05_02_preview.models import (
-    AgentPool,
     AgentPoolUpgradeSettings,
-    ContainerServiceStorageProfileTypes,
     CreationData,
     KubeletConfig,
     LinuxOSConfig,
@@ -2389,10 +2387,9 @@ def aks_pod_identity_add(cmd, client, resource_group_name, cluster_name,
         pod_identity.binding_selector = binding_selector
     pod_identities.append(pod_identity)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2419,10 +2416,9 @@ def aks_pod_identity_delete(cmd, client, resource_group_name, cluster_name,
                 continue
             pod_identities.append(pod_identity)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2452,10 +2448,9 @@ def aks_pod_identity_exception_add(cmd, client, resource_group_name, cluster_nam
         name=exc_name, namespace=exc_namespace, pod_labels=pod_labels)
     pod_identity_exceptions.append(exc)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2481,10 +2476,9 @@ def aks_pod_identity_exception_delete(cmd, client, resource_group_name, cluster_
                 continue
             pod_identity_exceptions.append(exc)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2518,10 +2512,9 @@ def aks_pod_identity_exception_update(cmd, client, resource_group_name, cluster_
         raise CLIError(
             'pod identity exception {}/{} not found'.format(exc_namespace, exc_name))
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,