+++ src/connectedk8s/azext_connectedk8s/_validators.py
@@ -2,6 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+import azext_connectedk8s._constants as consts
 
 
 def example_name_or_id_validator(cmd, namespace):
@@ -17,3 +18,10 @@ def example_name_or_id_validator(cmd, namespace):
                 type='storageAccounts',
                 name=namespace.storage_account
             )
+
+
+def override_client_request_id_header(cmd, namespace):
+    if namespace.correlation_id is not None:
+        cmd.cli_ctx.data['headers'][consts.Client_Request_Id_Header] = namespace.correlation_id
+    else:
+        cmd.cli_ctx.data['headers'][consts.Client_Request_Id_Header] = consts.Default_Onboarding_Source_Tracking_Guid