+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_scenario.py
@@ -85,6 +85,7 @@ class ContainerappScenarioTest(ScenarioTest):
         ])
 
 
+    # TODO rename
     @AllowLargeResponse(8192)
     @ResourceGroupPreparer(location="eastus2")
     def test_container_acr(self, resource_group):
@@ -126,6 +127,7 @@ class ContainerappScenarioTest(ScenarioTest):
         ])
 
 
+    # TODO rename
     @AllowLargeResponse(8192)
     @ResourceGroupPreparer(location="westeurope")
     def test_containerapp_update(self, resource_group):
@@ -193,6 +195,7 @@ class ContainerappScenarioTest(ScenarioTest):
         ])
 
 
+    # TODO rename
     @AllowLargeResponse(8192)
     @ResourceGroupPreparer(location="eastus2")
     def test_container_acr(self, resource_group):
@@ -258,6 +261,7 @@ class ContainerappScenarioTest(ScenarioTest):
             # Removing ACR password should fail since it is needed for ACR
             self.cmd('containerapp secret remove -g {} -n {} --secret-names {}'.format(resource_group, containerapp_name, secret_name))
 
+    # TODO rename
     @AllowLargeResponse(8192)
     @ResourceGroupPreparer(location="northeurope")
     def test_containerapp_update(self, resource_group):
@@ -329,6 +333,7 @@ class ContainerappScenarioTest(ScenarioTest):
             JMESPathCheck('properties.template.containers[1].resources.memory', '1.5Gi'),
         ])
 
+    # TODO fix and enable
     @unittest.skip("API only on stage currently")
     @live_only()  # VCR.py can't seem to handle websockets (only --live works)
     # @ResourceGroupPreparer(location="centraluseuap")
@@ -392,7 +397,6 @@ class ContainerappScenarioTest(ScenarioTest):
         for line in expected_output:
             self.assertIn(line, expected_output)
 
-
     @ResourceGroupPreparer(location="northeurope")
     def test_containerapp_logstream(self, resource_group):
         containerapp_name = self.create_random_name(prefix='capp', length=24)
@@ -412,3 +416,38 @@ class ContainerappScenarioTest(ScenarioTest):
         self.cmd(f'containerapp create -g {resource_group} -n {containerapp_name} --environment {env_name} --min-replicas 1 --ingress external --target-port 80')
 
         self.cmd(f'containerapp logs show -n {containerapp_name} -g {resource_group}')
+
+    @ResourceGroupPreparer(location="northeurope")
+    def test_containerapp_registry_msi(self, resource_group):
+        env = self.create_random_name(prefix='env', length=24)
+        logs = self.create_random_name(prefix='logs', length=24)
+        app = self.create_random_name(prefix='app', length=24)
+        acr = self.create_random_name(prefix='acr', length=24)
+
+        logs_id = self.cmd(f"monitor log-analytics workspace create -g {resource_group} -n {logs}").get_output_in_json()["customerId"]
+        logs_key = self.cmd(f'monitor log-analytics workspace get-shared-keys -g {resource_group} -n {logs}').get_output_in_json()["primarySharedKey"]
+
+        self.cmd(f'containerapp env create -g {resource_group} -n {env} --logs-workspace-id {logs_id} --logs-workspace-key {logs_key}')
+
+        containerapp_env = self.cmd(f'containerapp env show -g {resource_group} -n {env}').get_output_in_json()
+
+        while containerapp_env["properties"]["provisioningState"].lower() == "waiting":
+            time.sleep(5)
+            containerapp_env = self.cmd(f'containerapp env show -g {resource_group} -n {env}').get_output_in_json()
+
+        self.cmd(f'containerapp create -g {resource_group} -n {app} --environment {env} --min-replicas 1 --ingress external --target-port 80')
+        self.cmd(f'acr create -g {resource_group} -n {acr} --sku basic --admin-enabled')
+        # self.cmd(f'acr credential renew -n {acr} ')
+        self.cmd(f'containerapp registry set --server {acr}.azurecr.io -g {resource_group} -n {app}')
+        app_data = self.cmd(f'containerapp show -g {resource_group} -n {app}').get_output_in_json()
+        self.assertEqual(app_data["properties"]["configuration"]["registries"][0].get("server"), f'{acr}.azurecr.io')
+        self.assertIsNotNone(app_data["properties"]["configuration"]["registries"][0].get("passwordSecretRef"))
+        self.assertIsNotNone(app_data["properties"]["configuration"]["registries"][0].get("username"))
+        self.assertIsNone(app_data["properties"]["configuration"]["registries"][0].get("identity"))
+
+        self.cmd(f'containerapp registry set --server {acr}.azurecr.io -g {resource_group} -n {app} --identity system')
+        app_data = self.cmd(f'containerapp show -g {resource_group} -n {app}').get_output_in_json()
+        self.assertEqual(app_data["properties"]["configuration"]["registries"][0].get("server"), f'{acr}.azurecr.io')
+        self.assertIsNone(app_data["properties"]["configuration"]["registries"][0].get("passwordSecretRef"))
+        self.assertIsNone(app_data["properties"]["configuration"]["registries"][0].get("username"))
+        self.assertEqual(app_data["properties"]["configuration"]["registries"][0].get("identity"), "system")